Move the "fe" module into the "deneb-fuse" sub-crate
