+++ Cargo.toml
@@ -5,7 +5,7 @@ authors = ["Radu Popescu <mail@radupopescu.net>"]
 
 [features]
 default = ["fuse_module"]
-fuse_module = ["fuse"]
+fuse_module = ["deneb-fuse"]
 
 [[bin]]
 name = "deneb"
@@ -16,21 +16,13 @@ required-features = ["fuse_module"]
 [dependencies]
 deneb-common = { path = "deneb-common" }
 deneb-core = { path = "deneb-core" }
+deneb-fuse = { path = "deneb-fuse", optional = true }
+
 failure = "^0.1.0"
 log = "^0.4.0"
 log4rs = "^0.8.0"
 nix = "^0.9.0"
 structopt = "0.1.0"
 structopt-derive = "0.1.0"
-time = "^0.1.0"
-
-fuse = { version = "^0.3.0", optional = true}
-
-[dev-dependencies]
-copy_dir = "^0.1.0"
-quickcheck = "^0.6.0"
-rand = "^0.4.0"
-tempdir = "^0.3.0"
-uuid = { version = "^0.5.0", features = ["v4"]}
 
 [workspace]
