+++ deneb-fuse/src/lib.rs
@@ -1,7 +1,8 @@
-#[cfg(test)]
-mod tests {
-    #[test]
-    fn it_works() {
-        assert_eq!(2 + 2, 4);
-    }
-}
+extern crate fuse;
+extern crate nix;
+extern crate time;
+
+extern crate deneb_common;
+extern crate deneb_core;
+
+pub mod fs;
