--- src/fe/fuse/mod.rs
@@ -1,5 +1,0 @@
-//! Modules which make up the FUSE front-end
-
-mod fs;
-
-pub use self::fs::*;
