--- src/fe/mod.rs
@@ -1,4 +1,0 @@
-//! Frond-end modules of the application
-
-#[cfg(feature = "fuse")]
-pub mod fuse;
