+++ src/lib.rs
@@ -1,19 +1,17 @@
 #![recursion_limit = "1024"]
 
 extern crate failure;
-#[cfg(feature = "fuse")] extern crate fuse;
 extern crate log;
 extern crate log4rs;
 extern crate nix;
 extern crate structopt;
 #[macro_use] extern crate structopt_derive;
-extern crate time;
 
 // Crates from the workspace
 extern crate deneb_common;
 extern crate deneb_core;
+#[cfg(feature = "fuse_module")] extern crate deneb_fuse;
 
-pub mod fe;
 pub mod logging;
 pub mod params;
 pub mod util;
